Move spinner to libs/ui/
